+++ ui/shared/components/panel/family/Family.jsx
@@ -4,12 +4,13 @@ import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
 
 import { updateFamily } from 'redux/rootReducer'
-import { getProjectsByGuid } from 'redux/selectors'
+import { getProjectsByGuid, getNotesByFamilyType } from 'redux/selectors'
 
 import PedigreeImagePanel from '../view-pedigree-image/PedigreeImagePanel'
 import BaseFieldView from '../view-fields/BaseFieldView'
 import OptionFieldView from '../view-fields/OptionFieldView'
 import ListFieldView from '../view-fields/ListFieldView'
+import NoteListFieldView from '../view-fields/NoteListFieldView'
 import SingleFieldView from '../view-fields/SingleFieldView'
 import TagFieldView from '../view-fields/TagFieldView'
 import TextFieldView from '../view-fields/TextFieldView'
@@ -25,15 +26,13 @@ import {
   FAMILY_FIELD_FIRST_SAMPLE,
   FAMILY_FIELD_NAME_LOOKUP,
   FAMILY_FIELD_OMIM_NUMBER,
-  FAMILY_FIELD_PMIDS, FAMILY_FIELD_DESCRIPTION, FAMILY_FIELD_SUCCESS_STORY, FAMILY_FIELD_ANALYSIS_NOTES,
-  FAMILY_FIELD_ANALYSIS_SUMMARY, FAMILY_FIELD_MME_NOTES, FAMILY_FIELD_CODED_PHENOTYPE, FAMILY_FIELD_INTERNAL_NOTES,
-  FAMILY_FIELD_INTERNAL_SUMMARY,
+  FAMILY_FIELD_PMIDS, FAMILY_FIELD_DESCRIPTION, FAMILY_FIELD_SUCCESS_STORY, FAMILY_NOTES_FIELDS,
+  FAMILY_FIELD_CODED_PHENOTYPE, FAMILY_FIELD_INTERNAL_NOTES, FAMILY_FIELD_INTERNAL_SUMMARY,
 } from '../../../utils/constants'
 import { FirstSample, AnalystEmailDropdown, AnalysedBy, analysisStatusIcon } from './FamilyFields'
 import FamilyLayout from './FamilyLayout'
 
-
-const EDIT_FIELDS = [
+const ASSIGNED_ANALYST_EDIT_FIELDS = [
   {
     name: 'assigned_analyst_username',
     label: 'Email',
@@ -43,6 +42,20 @@ const EDIT_FIELDS = [
   },
 ]
 
+const mapNotesStateToProps = (state, ownProps) => ({
+  notes: (getNotesByFamilyType(state)[ownProps.initialValues.familyGuid] || {})[ownProps.modalId],
+})
+
+const BASE_NOTE_FIELD = {
+  canEdit: true,
+  component: connect(mapNotesStateToProps)(NoteListFieldView),
+}
+
+const getNoteField = noteType => ({
+  modalId: noteType,
+  submitArgs: { noteType, nestedField: 'note' },
+  ...BASE_NOTE_FIELD,
+})
 
 const FAMILY_FIELD_RENDER_LOOKUP = {
   [FAMILY_FIELD_DESCRIPTION]: { canEdit: true },
@@ -54,7 +67,7 @@ const FAMILY_FIELD_RENDER_LOOKUP = {
   },
   [FAMILY_FIELD_ASSIGNED_ANALYST]: {
     canEdit: true,
-    formFields: EDIT_FIELDS,
+    formFields: ASSIGNED_ANALYST_EDIT_FIELDS,
     component: BaseFieldView,
     submitArgs: { familyField: 'assigned_analyst' },
     addConfirm: 'Are you sure you want to add the analyst to this family?',
@@ -81,9 +94,6 @@ const FAMILY_FIELD_RENDER_LOOKUP = {
     fieldDisplay: (loadedSample, compact, familyGuid) =>
       <FirstSample familyGuid={familyGuid} compact={compact} />,
   },
-  [FAMILY_FIELD_ANALYSIS_NOTES]: { canEdit: true },
-  [FAMILY_FIELD_ANALYSIS_SUMMARY]: { canEdit: true },
-  [FAMILY_FIELD_MME_NOTES]: { canEdit: true },
   [FAMILY_FIELD_CODED_PHENOTYPE]: { component: SingleFieldView, canEdit: true },
   [FAMILY_FIELD_OMIM_NUMBER]: {
     canEdit: true,
@@ -99,6 +109,7 @@ const FAMILY_FIELD_RENDER_LOOKUP = {
   },
   [FAMILY_FIELD_INTERNAL_NOTES]: { internal: true, submitArgs: { familyField: 'case_review_notes' } },
   [FAMILY_FIELD_INTERNAL_SUMMARY]: { internal: true, submitArgs: { familyField: 'case_review_summary' } },
+  ...FAMILY_NOTES_FIELDS.reduce((acc, { id, noteType }) => ({ ...acc, [id]: getNoteField(noteType) }), {}),
 }
 
 const Family = React.memo((